♻️ refactor(components): separate `EndpointIcon` from `Icon` component for standalone use
